Remove escape sequences for newline and tab
- Simpler parsing
- More consistent (these were special shorthands)
- Leaves open the possibility for escaping names

The tradeoff is that it takes more characters to escape newline `\00000A` and tab `\000009`. It would be possible to add shorthands, eg. `\!n` and `\!t` or `\(A)` and `\(9)` (no zero-padding). This is left as a possibility for supersets derived from na. Any character that's not allowed as the first character of a name is still eligible for escape sequence use.